cleaned up structure assertions
